More refactoring, integration test still not done
