--- policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/ctc_arpa_reduction.yaml
@@ -1,41 +1,0 @@
-- name: ARPA reduction only applies for 2021.
-  period: 2020
-  input:
-    adjusted_gross_income: 220_000
-  output:
-    ctc_arpa_reduction: 0
-
-- name: Head of household with $220k income and a child under six has a $2,600 CTC reduction under ARPA.
-  period: 2021
-  input:
-    adjusted_gross_income: 220_000
-    ctc_phase_out_threshold: 200_000 # HOH.
-    ctc_maximum: 3_600 # Young child amount.
-    filing_status: HEAD_OF_HOUSEHOLD
-    ctc_maximum_without_arpa: 2_000
-  output:
-    ctc_arpa_reduction: 1_600 # Full $1,600 additional CTC.
-
-- name: Joint filer with $160k income has a phase-out of $500 due to ARPA. 
-  period: 2021
-  input:
-    adjusted_gross_income: 160_000
-    ctc_phase_out_threshold: 200_000 # HOH.
-    ctc_maximum: 3_000 # Non-young child amount.
-    filing_status: JOINT
-    ctc_maximum_without_arpa: 2_000
-  output:
-    ctc_arpa_reduction: 500 # $10k excess above $150k ARPA threshold * 5%.
-
-- name: 10 kids, single parent, $210k income.
-  period: 2021
-  input: 
-    adjusted_gross_income: 210_000
-    ctc_phase_out_threshold: 200_000 # HOH.
-    ctc_maximum: 30_000 # 10 * non-young child amount.
-    filing_status: HEAD_OF_HOUSEHOLD
-    ctc_maximum_without_arpa: 20_000
-  output:
-    # Difference in phase-out range is $87,500.
-    # 5% is $4,375.
-    ctc_reduction: 500 # $10k excess above $150k ARPA threshold * 5%.
