+++ policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/maximum/individual/ctc_child_individual_maximum_arpa.yaml
@@ -1,0 +1,13 @@
+- name: ARP provided $3,000 to non-young children.
+  period: 2021
+  input:
+    age: 12
+  output:
+    ctc_child_individual_maximum_arpa: 3_000
+
+- name: ARP provided $3,600 to young children.
+  period: 2021
+  input:
+    age: 5
+  output:
+    ctc_child_individual_maximum_arpa: 3_600
