--- policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/maximum/individual/is_ctc_qualifying_child.yaml
@@ -1,34 +1,0 @@
-- name: 16-year-olds qualified prior to 2021.
-  period: 2020
-  input:
-    age: 16
-  output:
-    is_ctc_qualifying_child: true
-
-- name: 17-year-olds did not qualify prior to 2021.
-  period: 2020
-  input:
-    age: 17
-  output:
-    is_ctc_qualifying_child: false
-
-- name: 17-year-olds qualified in 2021.
-  period: 2021
-  input:
-    age: 17
-  output:
-    is_ctc_qualifying_child: true
-
-- name: 18-year-olds did not qualify in 2021.
-  period: 2021
-  input:
-    age: 18
-  output:
-    is_ctc_qualifying_child: false
-
-- name: 17-year-olds did not qualify after 2021.
-  period: 2022
-  input:
-    age: 17
-  output:
-    is_ctc_qualifying_child: false
