--- policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/maximum/individual/is_ctc_qualifying_young_child.yaml
@@ -1,20 +1,0 @@
-- name: Young children are those under age six, as defined by ARPA in 2021.
-  period: 2021
-  input:
-    age: 5
-  output:
-    is_ctc_qualifying_young_child: true
-
-- name: Six-year-olds aren't.
-  period: 2022
-  input:
-    age: 6
-  output:
-    is_ctc_qualifying_young_child: false
-
-- name: Not defined before 2021.
-  period: 2020
-  input:
-    age: 5
-  output:
-    is_ctc_qualifying_young_child: false
