+++ policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/reduction/arpa/ctc_arpa_addition.yaml
@@ -1,0 +1,7 @@
+- name: Minimum of max addition and reduction.
+  period: 2020
+  input:
+    ctc_arpa_max_addition: 1
+    ctc_arpa_reduction: 2
+  output:
+    ctc_arpa_addition: 1
