+++ policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/reduction/arpa/ctc_arpa_addition_cap.yaml
@@ -1,0 +1,7 @@
+- name: 5% of difference between ARPA and non-ARPA phase-out thresholds (joint).
+  period: 2021
+  input:
+    ctc_arpa_phase_out_threshold: 150_000
+    ctc_phase_out_threshold: 400_000
+  output:
+    ctc_arpa_addition_cap: 12_500 # 400k - 150k = 250k * 0.05 = 12.5k
