+++ policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/reduction/ctc_phase_out_threshold.yaml
